Разбить JS код на модули
