+++ js/const.js
@@ -1,0 +1,50 @@
+const NAMES = [
+  'Даниил Покровский',
+  'Любовь Смирнова',
+  'Анна Бычкова',
+  'Вероника Игнатова',
+  'Ульяна Горбачева',
+  'Максим Морозов',
+  'Лилия Гаврилова',
+  'Сергей Денисов',
+  'Варвара Григорьева',
+  'Арсений Горшков',
+  'Богдан Малышев',
+  'Артём Иванов',
+  'Денис Макеев',
+  'Софья Орлова',
+  'Андрей Егоров',
+  'Елизавета Журавлева',
+  'Максим Михеев',
+  'Макар Покровский',
+  'Артём Дмитриев',
+  'Анастасия Крылова',
+  'Максим Иванов',
+  'Серафима Александрова',
+  'Вячеслав Николаев',
+  'Роман Марков',
+  'Лидия Яковлева'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTIONS = [
+  'wow, wow!',
+  'Loooove',
+  '😍😍😍😍',
+  'holy shit!',
+  'Nice👍🏻',
+  'magic',
+  'very crazey',
+  'Mamma mia',
+  'Beautiful 🙏🏼'
+];
+
+export {NAMES, MESSAGES, DESCRIPTIONS};
